requirements.txt
# Core dependencies
kubernetes==28.1.0
docker==7.0.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Scheduling
APScheduler==3.10.4
croniter==2.0.1

# API and Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Notifications
requests==2.31.0
slack-sdk==3.26.2

# Monitoring and Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# CLI
click==8.1.7
tabulate==0.9.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1
aiohttp>=3.9.0
